+++ bin/pyrdp-mitm.py
@@ -17,48 +17,18 @@ from pyrdp.core.ssl import ServerTLSContext
 asyncioreactor.install(asyncio.get_event_loop())
 
 import argparse
-import logging
 import logging.handlers
 import os
-import random
 import sys
 from pathlib import Path
 
 import appdirs
-import names
 from twisted.internet import reactor
-from twisted.internet.protocol import ServerFactory
 
 from pyrdp.core import parseTarget
+from pyrdp.core.mitm import MITMServerFactory
 from pyrdp.logging import JSONFormatter, log, LOGGER_NAMES, LoggerNameFilter, SessionLogger, VariableFormatter
 from pyrdp.mitm import MITMConfig, RDPMITM
-
-
-class MITMServerFactory(ServerFactory):
-    """
-    Server factory for the RDP man-in-the-middle that generates a unique session ID for every connection.
-    """
-
-    def __init__(self, config: MITMConfig):
-        """
-        :param config: the MITM configuration
-        """
-        self.config = config
-
-    def buildProtocol(self, addr):
-        sessionID = f"{names.get_first_name()}{random.randrange(100000,999999)}"
-
-        # mainLogger logs in a file and stdout
-        mainlogger = logging.getLogger(LOGGER_NAMES.MITM_CONNECTIONS)
-        mainlogger = SessionLogger(mainlogger, sessionID)
-
-        # crawler logger only logs to a file for analysis purposes
-        crawlerLogger = logging.getLogger(LOGGER_NAMES.CRAWLER)
-        crawlerLogger = SessionLogger(crawlerLogger, sessionID)
-
-        mitm = RDPMITM(mainlogger, crawlerLogger, self.config)
-
-        return mitm.getProtocol()
 
 
 def prepareLoggers(logLevel: int, logFilter: str, sensorID: str, outDir: Path):
@@ -173,6 +143,7 @@ def logConfiguration(config: MITMConfig):
 
 
 def main():
+    # Warning: keep in sync with twisted/plugins/pyrdp_plugin.py
     parser = argparse.ArgumentParser()
     parser.add_argument("target", help="IP:port of the target RDP machine (ex: 192.168.1.10:3390)")
     parser.add_argument("-l", "--listen", help="Port number to listen on (default: 3389)", default=3389)
